Moved covariance components out of their respective sections
