+++ cmake/unit_testing.cmake
@@ -33,6 +33,9 @@ add_subdirectory( src/ENDFtk/record/Sequence/test )
 add_subdirectory( src/ENDFtk/record/Tail/test )
 add_subdirectory( src/ENDFtk/record/Zipper/test )
 add_subdirectory( src/ENDFtk/record/test )
+add_subdirectory( src/ENDFtk/section/CovariancePairs/test )
+add_subdirectory( src/ENDFtk/section/RectangularMatrix/test )
+add_subdirectory( src/ENDFtk/section/SquareMatrix/test )
 add_subdirectory( src/ENDFtk/section/1/451/test )
 add_subdirectory( src/ENDFtk/section/1/452/test )
 add_subdirectory( src/ENDFtk/section/1/455/DecayConstants/test )
@@ -123,18 +126,12 @@ add_subdirectory( src/ENDFtk/section/32/151/BreitWignerLValue/Resonance/test )
 add_subdirectory( src/ENDFtk/section/32/151/BreitWignerLValue/test )
 add_subdirectory( src/ENDFtk/section/32/151/LimitedSingleLevelBreitWigner/test )
 add_subdirectory( src/ENDFtk/section/32/151/LimitedMultiLevelBreitWigner/test )
-add_subdirectory( src/ENDFtk/section/33/CovariancePairs/test )
 add_subdirectory( src/ENDFtk/section/33/DerivedRatioToStandard/test )
 add_subdirectory( src/ENDFtk/section/33/DerivedRedundant/test )
 add_subdirectory( src/ENDFtk/section/33/ReactionBlock/test )
-add_subdirectory( src/ENDFtk/section/33/RectangularMatrix/test )
-add_subdirectory( src/ENDFtk/section/33/SquareMatrix/test )
 add_subdirectory( src/ENDFtk/section/33/test )
-add_subdirectory( src/ENDFtk/section/34/CovariancePairs/test )
 add_subdirectory( src/ENDFtk/section/34/LegendreBlock/test )
 add_subdirectory( src/ENDFtk/section/34/ReactionBlock/test )
-add_subdirectory( src/ENDFtk/section/34/RectangularMatrix/test )
-add_subdirectory( src/ENDFtk/section/34/SquareMatrix/test )
 add_subdirectory( src/ENDFtk/section/34/test )
 add_subdirectory( src/ENDFtk/section/4/Isotropic/test )
 add_subdirectory( src/ENDFtk/section/4/LegendreCoefficients/test )
