+++ src/ENDFtk/section/33.hpp
@@ -16,6 +16,10 @@
 #include "ENDFtk/readSequence.hpp"
 #include "ENDFtk/section.hpp"
 
+#include "ENDFtk/section/CovariancePairs.hpp"
+#include "ENDFtk/section/SquareMatrix.hpp"
+#include "ENDFtk/section/RectangularMatrix.hpp"
+
 namespace njoy {
 namespace ENDFtk {
 namespace section{
@@ -38,9 +42,6 @@ namespace section{
       DerivedRedundant,
       DerivedRatioToStandard >;
 
-    #include "ENDFtk/section/33/CovariancePairs.hpp"
-    #include "ENDFtk/section/33/SquareMatrix.hpp"
-    #include "ENDFtk/section/33/RectangularMatrix.hpp"
     using ExplicitCovariance = std::variant<
       CovariancePairs,
       SquareMatrix,
