+++ src/ENDFtk/section/33/test/33.test.cpp
@@ -13,9 +13,9 @@ using DerivedCovariance = section::Type< 33 >::DerivedCovariance;
 using ExplicitCovariance = section::Type< 33 >::ExplicitCovariance;
 using DerivedRedundant = section::Type< 33 >::DerivedRedundant;
 using DerivedRatioToStandard = section::Type< 33 >::DerivedRatioToStandard;
-using SquareMatrix = section::Type< 33 >::SquareMatrix;
-using RectangularMatrix = section::Type< 33 >::RectangularMatrix;
-using CovariancePairs = section::Type< 33 >::CovariancePairs;
+using SquareMatrix = section::SquareMatrix;
+using RectangularMatrix = section::RectangularMatrix;
+using CovariancePairs = section::CovariancePairs;
 
 std::string chunk();
 void verifyChunk( const section::Type< 33 >& );
