+++ src/ENDFtk/section/34/ReactionBlock/test/ReactionBlock.test.cpp
@@ -9,7 +9,7 @@
 using namespace njoy::ENDFtk;
 using ReactionBlock = section::Type< 34 >::ReactionBlock;
 using LegendreBlock = section::Type< 34 >::LegendreBlock;
-using SquareMatrix = section::Type< 34 >::SquareMatrix;
+using SquareMatrix = section::SquareMatrix;
 
 std::string chunk();
 void verifyChunk( const ReactionBlock& );
