+++ src/ENDFtk/section/RectangularMatrix/src/ctor.hpp
