+++ src/ENDFtk/section/SquareMatrix/src/ctor.hpp
